LayoutTests/fast/viewport/extrazoom/viewport-change-min-device-width.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<html>
<meta name="viewport" id="meta">
<head>
    <script src="../../../resources/js-test.js"></script>
    <script src="../../../resources/ui-helper.js"></script>
    <style>
        body, html {
            margin: 0;
            width: 100%;
            height: 100%;
        }
    </style>
    <script>
        jsTestIsAsync = true;

        function logWindowDimensionsAfterSettingViewportContent(content) {
            return new Promise(async resolve => {
                meta.setAttribute("content", content);
                await UIHelper.ensurePresentationUpdate();
                debug(`[${meta.getAttribute("content")}] (${innerWidth}, ${innerHeight})`);
                resolve();
            });
        }

        async function runTest() {
            if (!window.testRunner) {
                description("Please use WebKitTestRunner to run this test.");
                return;
            }

            debug("1. Default viewport");
            await logWindowDimensionsAfterSettingViewportContent("width=150");
            await logWindowDimensionsAfterSettingViewportContent("width=device-width");
            scaleAtDeviceWidthWithInitialShrinkToFit = parseFloat(await UIHelper.zoomScale()).toFixed(3);
            await logWindowDimensionsAfterSettingViewportContent("width=600");

            debug("\n2. shrink-to-fit explicitly disabled");
            await logWindowDimensionsAfterSettingViewportContent("width=150, shrink-to-fit=no");
            await logWindowDimensionsAfterSettingViewportContent("width=device-width, shrink-to-fit=0");
            scaleAtDeviceWidthWithShrinkToFitDisabled = parseFloat(await UIHelper.zoomScale()).toFixed(3);
            await logWindowDimensionsAfterSettingViewportContent("width=600, shrink-to-fit=-0.5");

            debug("\n3. shrink-to-fit explicitly enabled");
            await logWindowDimensionsAfterSettingViewportContent("width=150, shrink-to-fit=yes");
            await logWindowDimensionsAfterSettingViewportContent("width=device-width, shrink-to-fit=1");
            scaleAtDeviceWidthWithShrinkToFitEnabled = parseFloat(await UIHelper.zoomScale()).toFixed(3);
            await logWindowDimensionsAfterSettingViewportContent("width=600, shrink-to-fit=device-width");

            shouldBe("scaleAtDeviceWidthWithInitialShrinkToFit", "'0.488'");
            shouldBe("scaleAtDeviceWidthWithShrinkToFitDisabled", "'1.000'");
            shouldBe("scaleAtDeviceWidthWithShrinkToFitEnabled", "'0.488'");

            finishJSTest();
        }
    </script>
</head>
<body onload="runTest()">
</body>
</html>
